+++ Tests/Editor/LookupExtensionsTests.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using NUnit.Framework;
+using UnityEngine;
+using static Tutan.Functional.F;
 
 namespace Tutan.Functional.Tests
 {
@@ -73,6 +75,43 @@ namespace Tutan.Functional.Tests
             Optional<string> result = dict.Lookup("key");
 
             Assert.IsTrue(result.IsNone);
+        }
+
+        [Test]
+        public void Alive_LiveObject_ReturnsSome()
+        {
+            var go = new GameObject("alive-test");
+            try
+            {
+                Optional<GameObject> opt = Some(go);
+
+                Assert.IsTrue(opt.Alive().IsSome);
+            }
+            finally
+            {
+                Object.DestroyImmediate(go);
+            }
+        }
+
+        [Test]
+        public void Alive_DestroyedObject_ReturnsNone()
+        {
+            var go = new GameObject("alive-test");
+            Optional<GameObject> opt = Some(go);
+            Object.DestroyImmediate(go);
+
+            // The snapshot taken at Some() is still stale-Some...
+            Assert.IsTrue(opt.IsSome);
+            // ...but Alive() re-checks Unity lifetime.
+            Assert.IsTrue(opt.Alive().IsNone);
+        }
+
+        [Test]
+        public void Alive_None_ReturnsNone()
+        {
+            Optional<GameObject> opt = None;
+
+            Assert.IsTrue(opt.Alive().IsNone);
         }
     }
 }